Add initial Mbed CE support for Ambiq Apollo3 (#469)
* Upgrade Ambiq Apollo3 linker script

* Add PYOCD, fix conflict between RAM and RAM vectors that was making boot fail

* Add pre-sync delay, fixing inability for greentea to work on Ambiq Apollo3

* Start on sync predelay option

* Tests passing on Artemis! There was some missing undocumented lp ticker behavior, I added docs and tests for it.

* Add J-Link

* Typo

* Fix some CI issues

+++ connectivity/drivers/ble/FEATURE_BLE/TARGET_Ambiq_Micro/hal/apollo3/hci_drv_apollo3.c
@@ -511,7 +511,7 @@ HciDrvRadioBoot(bool bColdBoot)
     CRITICAL_PRINT("INTEN:  %d\n", BLEIF->INTEN_b.BLECSSTAT);
     CRITICAL_PRINT("INTENREG:  %d\n", BLEIF->INTEN);
 
-    NVIC_SetVector(BLE_IRQn, HciDrvIntService);
+    NVIC_SetVector(BLE_IRQn, (uint32_t)HciDrvIntService);
     NVIC_EnableIRQ(BLE_IRQn);
 
     //
